Move HelperClass and external_callback to concurrent for testing
